requirements.txt
openai>=1.0.0
pillow>=10.0.0
imagehash>=4.3.0
opencv-python-headless>=4.8.0
tqdm>=4.65.0
python-dotenv>=1.0.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
umap-learn>=0.5.3
hdbscan>=0.8.33
face-recognition>=1.3.0
dlib>=19.24.0
